+++ cf/read_write/read.py
@@ -1,5 +1,6 @@
 import logging
 import os
+import tempfile
 
 from glob import glob
 from os.path import isdir
@@ -26,6 +27,8 @@ else:
     python36 = False
 
 
+_cached_temporary_files = {}
+
 # --------------------------------------------------------------------
 # Create an implementation container and initialize a read object for
 # each format
@@ -50,6 +53,7 @@ def read(
     squeeze=False,
     unsqueeze=False,
     fmt=None,
+    cdl_string=False,
     select=None,
     extra=None,
     recursive=False,
@@ -207,6 +211,11 @@ def read(
             be raised, unless the *ignore_read_error* parameter is
             True.
 
+            As a special case, if the `cdl_string` parameter is set to
+            True, the interpretation of `files` changes so that each
+            value is assumed to be a string of CDL input rather
+            than the above.
+
         external: (sequence of) `str`, optional
             Read external variables (i.e. variables which are named by
             attributes, but are not present, in the parent file given
@@ -312,6 +321,22 @@ def read(
             ``'CFA'`` for CFA-netCDF files, ``'UM'`` for PP or UM
             fields files, and ``'CDL'`` for CDL text files. By default
             files of any of these formats are read.
+
+        cdl_string: `bool`, optional
+            If True and the format to read is CDL, read a string
+            input, or sequence of string inputs, each being interpreted
+            as a string of CDL rather than names of locations from
+            which field constructs can be read from, as standard.
+
+            By default, each string input or string element in the input
+            sequence is taken to be a file or directory name or an
+            OPenDAP URL from which to read field constructs, rather
+            than a string of CDL input, including when the `fmt`
+            parameter is set as CDL.
+
+            Note that when `cdl_string` is True, the `fmt` parameter is
+            ignored as the format is assumed to be CDL, so in that case
+            it is not necessary to also specify ``fmt='CDL'``.
 
         aggregate: `bool` or `dict`, optional
             If True (the default) or a dictionary (possibly empty)
@@ -576,9 +601,20 @@ def read(
         if isinstance(select, (str, Query, Pattern)):
             select = (select,)
 
+    # Manage input parameters where contradictions are possible:
+    if cdl_string and fmt:
+        if fmt == "CDL":
+            logger.info(
+                "It is not necessary to set the cf.read fmt as 'CDL' when "
+                "cdl_string is True, since that implies CDL is the format."
+            )  # pragma: no cover
+        else:
+            raise ValueError(
+                "cdl_string can only be True when the format is CDL, though "
+                "fmt is ignored in that case so there is no need to set it."
+            )
     if squeeze and unsqueeze:
         raise ValueError("squeeze and unsqueeze can not both be True")
-
     if follow_symlinks and not recursive:
         raise ValueError(
             "Can't set follow_symlinks={0} when recursive={1}".format(
@@ -609,6 +645,36 @@ def read(
     # files
     field_counter = -1
     file_counter = 0
+
+    if cdl_string:
+        files2 = []
+
+        # 'files' input may be a single string or a sequence of them and to
+        # handle both cases it is easiest to convert former to a one-item seq.
+        if isinstance(files, str):
+            files = [files]
+
+        for cdl_file in files:
+            c = tempfile.NamedTemporaryFile(
+                mode="w",
+                dir=tempfile.gettempdir(),
+                prefix="cf_",
+                suffix=".cdl",
+            )
+
+            c_name = c.name
+            with open(c_name, "w") as f:
+                f.write(cdl_file)
+
+            # ----------------------------------------------------------------
+            # Need to cache the TemporaryFile object so that it doesn't get
+            # deleted too soon
+            # ----------------------------------------------------------------
+            _cached_temporary_files[c_name] = c
+
+            files2.append(c.name)
+
+        files = files2
 
     for file_glob in flat(files):
         # Expand variables