+++ Changelog.rst
@@ -4,10 +4,12 @@ version 3.??.?
 
 **2021-??-??**
 
+* New keyword parameter to `cf.read`: ``cdl_string``
+  (https://github.com/NCAS-CMS/cf-python/issues/171)
 * Fix bug that causes a failure in `cf.aggregate` when otherwise
   aggregatable fields have non-valid units
   (https://github.com/NCAS-CMS/cf-python/issues/229)
-  
+
 version 3.10.0
 --------------
 ----